split timeseries in 4 files
