+++ modelskill/timeseries/__init__.py
@@ -1,4 +1,6 @@
-from ._timeseries import TimeSeries, _parse_point_input, _parse_track_input
+from ._timeseries import TimeSeries
+from ._point import _parse_point_input
+from ._track import _parse_track_input
 
 __all__ = [
     "TimeSeries",
